unified-model/pyproject.toml
[build-system]
requires = ["setuptools>=40.6.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tar"
version = "0.1.0"
description = "TAR: Multimodal Generative Model"
authors = [{ name = "Your Name", email = "your@email.com" }]
license = { file = "LICENSE-CODE" }
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "accelerate==0.28.0",
    "datasets==2.16.1",
    "deepspeed==0.14.4",
    "einops==0.8.1",
    "gradio==5.34.0",
    "huggingface_hub==0.29.1",
    "numpy==1.26.1",
    "Pillow==11.2.1",
    "pyarrow==17.0.0",
    "PyYAML==6.0.2",
    "torch==2.1.2",
    "torchvision==0.16.2",
    "tqdm==4.66.5",
    "transformers==4.50.0",
    "wandb",
    "easydict"
]

[tool.setuptools]
packages = {find = {exclude = ["images"]}}